scuffed version of dashboard page without redux
initial removal for dashboard

tewsting

testing again

fix: misc.

feat: use context method for toggling state

removed redux from package.json
