+++ .prettierrc
@@ -1,5 +1,5 @@
 {
-    "printWidth": 200,
+    "printWidth": 100,
     "tabWidth": 4,
     "useTabs": false,
     "semi": true,
