+++ gui/package.json
@@ -6,7 +6,6 @@
         "@emotion/react": "^11.11.3",
         "@emotion/styled": "^11.11.0",
         "@mui/material": "^5.15.10",
-        "@reduxjs/toolkit": "^2.2.0",
         "@testing-library/jest-dom": "^5.17.0",
         "@testing-library/react": "^13.4.0",
         "@testing-library/user-event": "^13.5.0",
@@ -18,7 +17,6 @@
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "react-gauge-component": "^1.2.0",
-        "react-redux": "^9.1.0",
         "react-router-dom": "^6.22.0",
         "react-scripts": "5.0.1",
         "styled-components": "^5.3.11",
