+++ gui/public/index.html
@@ -39,7 +39,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-        <title>React Redux App</title>
+        <title>Rocket GUI</title>
     </head>
     <body>
         <noscript>You need to enable JavaScript to run this app.</noscript>
