--- gui/src/redux/store.js
@@ -1,8 +1,0 @@
-import {configureStore} from "@reduxjs/toolkit";
-import rocketReducer from "./rocketSlice";
-
-export const store = configureStore({
-    reducer: {
-        rocket: rocketReducer
-    }
-});
