+++ webservice/server.py
@@ -66,7 +66,9 @@ gse_state = {
     "solenoidCurrentMvasOpen": 0,
     "solenoidCurrentMvasClose": 0,
     "solenoidCurrentGn2Disconnect": 0,
+    "solenoidCurrentLoxFill": 0,
     "solenoidCurrentLoxVent": 0,
+    "solenoidCurrentLngFill": 0,
     "solenoidCurrentLngVent": 0,
     "solenoidExpectedGn2Fill": -1,
     "solenoidExpectedGn2Vent": -1,
